+++ example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_bakong_khqr/core/khqr_curency.dart';
 import 'dart:async';
 
 import 'package:flutter_bakong_khqr/flutter_bakong_khqr.dart';
@@ -17,8 +18,6 @@ class FlutterBakongKhqrExample extends StatefulWidget {
 
 class _FlutterBakongKhqrExampleState extends State<FlutterBakongKhqrExample> {
   String _qrCode = '';
-  String _md5Hash = '';
-  String deeplink = '';
   final _bakongKhqr = FlutterBakongKhqr();
 
   @override
@@ -32,25 +31,14 @@ class _FlutterBakongKhqrExampleState extends State<FlutterBakongKhqrExample> {
     final response = await _bakongKhqr.generateKhqrIndividual(
       bakongAccountId: "sokha_tim@aclb",
       accountInformation: "855979515836",
-      acquiringBank: "Dev Bank",
-      amount: 10000,
+      amount: 100,
+      currency: KhqrCurrency.usd,
       merchantName: "Sokha Tim",
-      merchantCity: "PHNOM PENH",
-      billNumber: "#12345",
-      mobileNumber: "85512233455",
-      storeLabel: "Coffee Shop",
-      terminalLabel: "Cashier_1",
-      upiAccountInformation: "1234567812345678ABCDEFGHIJKLMNO",
-      purposeOfTransaction: "Buy coffee",
-      merchantAlternateLanguagePreference: "km",
-      merchantNameAlternateLanguage: "Sokha",
-      merchantCityAlternateLanguage: "Phnom Penh",
     );
 
     // Assuming response is a Map with keys 'qrCode' and 'md5'
     setState(() {
       _qrCode = response.qrCode;
-      _md5Hash = response.md5;
     });
     if (!mounted) return;
   }
@@ -68,37 +56,14 @@ class _FlutterBakongKhqrExampleState extends State<FlutterBakongKhqrExample> {
             ),
           ),
         ),
-        body: Column(
-          mainAxisSize: MainAxisSize.min,
-          children: [
-            const Text(
-              "KHQR String :",
-              style: TextStyle(fontSize: 20, fontWeight: FontWeight.w700),
-            ),
-            Text(
-              _qrCode,
-              // style: const TextStyle(fontSize: 20),
-            ),
-            const Divider(),
-            const Text(
-              "KHQR Image :",
-              style: TextStyle(fontSize: 20, fontWeight: FontWeight.w700),
-            ),
-            BakongKhqrView(
-              merchantName: "Sokha Tim",
-              qrString: _qrCode,
-              amount: '10000',
-            ),
-            const Divider(),
-            const Text(
-              "MD5 String :",
-              style: TextStyle(fontSize: 20, fontWeight: FontWeight.w700),
-            ),
-            Text(
-              _md5Hash,
-              style: const TextStyle(fontSize: 20),
-            ),
-          ],
+        body: Center(
+          child: BakongKhqrView(
+            width: 350,
+            amount: 100,
+            receiverName: "Sokha Tim",
+            currency: KhqrCurrency.usd,
+            qr: _qrCode,
+          ),
         ),
       ),
     );