+++ lib/core/khqr_curency.dart
@@ -1,0 +1,5 @@
+/// KHQR Currency Enum
+enum KhqrCurrency {
+  khr,
+  usd,
+}