+++ CHANGELOG.md
@@ -13,3 +13,9 @@
 ## 0.0.1+3
 
 * TODO: Fix bug.
+
+## 0.0.2
+
+* TODO: Fix bug.
+* Add USD Currency in Bakong Khqr View.
+* Can Generate QR Code with USD Currency.